+++ src/coq_elpi_HOAS.mli
@@ -11,6 +11,7 @@ open RawData
 type empty = [ `Options ]
 type full  = [ `Options | `Context ]
 
+type ppoption = All | Most | Normal
 type hole_mapping =
   | Verbatim   (* 1:1 correspondence between UVar and Evar *)
   | Heuristic  (* new UVar outside Llam is pruned before being linked to Evar *)
@@ -21,6 +22,8 @@ type options = {
   deprecation : Deprecation.t option;
   primitive : bool option;
   failsafe : bool; (* readback is resilient to illformed terms *)
+  ppwidth : int;
+  pp : ppoption;
 }
 
 type 'a coq_context = {