+++ src/coq_elpi_builtins.ml
@@ -24,6 +24,78 @@ open Names
 open Coq_elpi_utils
 open Coq_elpi_HOAS
 
+let string_of_ppcmds options pp =
+  let b = Buffer.create 512 in
+  let fmt = Format.formatter_of_buffer b in
+  Format.pp_set_margin fmt options.ppwidth;
+  Format.fprintf fmt "@[%a@]" Pp.pp_with pp;
+  Format.pp_print_flush fmt ();
+  Buffer.contents b
+
+let with_pp_options o f =
+  let raw_print = !Flags.raw_print in
+  let print_universes = !Constrextern.print_universes in
+  let print_no_symbol = !Constrextern.print_no_symbol in
+  (* 8.14 let print_primitive_token = !Constrextern.print_primitive_token in*)
+  let print_implicits = !Constrextern.print_implicits in
+  let print_coercions = !Constrextern.print_coercions in
+  let print_parentheses = !Constrextern.print_parentheses in
+  let print_projections = !Constrextern.print_projections in
+  let print_evar_arguments = !Constrextern.print_evar_arguments in
+  let f =
+    match o with
+    | All ->
+        Flags.raw_print := true;
+        f
+    | Most ->
+        Flags.raw_print := false;
+        Constrextern.print_universes := false;
+        Constrextern.print_no_symbol := true;
+        (* 8.14 Constrextern.print_primitive_token := true; *)
+        Constrextern.print_implicits := true;
+        Constrextern.print_coercions := true;
+        Constrextern.print_parentheses := true;
+        Constrextern.print_projections := false;
+        Constrextern.print_evar_arguments := false;
+        Constrextern.with_meta_as_hole f
+    | Normal ->
+        (* If no preference is given, we print using Coq's current value *)
+        f
+  in
+  try
+    let rc = f () in
+    Flags.raw_print := raw_print;
+    Constrextern.print_universes := print_universes;
+    Constrextern.print_no_symbol := print_no_symbol;
+    (* 8.14 Constrextern.print_primitive_token := print_primitive_token; *)
+    Constrextern.print_implicits := print_implicits;
+    Constrextern.print_coercions := print_coercions;
+    Constrextern.print_parentheses := print_parentheses;
+    Constrextern.print_projections := print_projections;
+    Constrextern.print_evar_arguments := print_evar_arguments;
+    rc
+  with reraise ->
+    Flags.raw_print := raw_print;
+    Constrextern.print_universes := print_universes;
+    Constrextern.print_no_symbol := print_no_symbol;
+    (* 8.14 Constrextern.print_primitive_token := print_primitive_token; *)
+    Constrextern.print_implicits := print_implicits;
+    Constrextern.print_coercions := print_coercions;
+    Constrextern.print_parentheses := print_parentheses;
+    Constrextern.print_projections := print_projections;
+    Constrextern.print_evar_arguments := print_evar_arguments;
+    raise reraise
+
+let pr_econstr_env options env sigma t =
+  with_pp_options options.pp (fun () ->
+    let expr = Constrextern.extern_constr env sigma t in
+    let expr =
+      let rec aux () ({ CAst.v } as orig) = match v with
+      | Constrexpr.CEvar _ -> CAst.make @@ Constrexpr.CHole(None,Namegen.IntroAnonymous,None)
+      | _ -> Constrexpr_ops.map_constr_expr_with_binders (fun _ () -> ()) aux () orig in
+      if options.hoas_holes = Some Heuristic then aux () expr else expr in
+    Ppconstr.pr_constr_expr env sigma expr)
+
 let tactic_mode = ref false
 let on_global_state api thunk = (); (fun state ->
   if !tactic_mode then
@@ -172,14 +244,14 @@ let univ = { univ with
 let term = {
   CConv.ty = Conv.TyName "term";
   pp_doc = (fun fmt () -> Format.fprintf fmt "A Coq term containing evars");
-  pp = (fun fmt t -> Format.fprintf fmt "%s" (Pp.string_of_ppcmds (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
+  pp = (fun fmt t -> Format.fprintf fmt "@[%a@]" Pp.pp_with ( (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
   readback = lp2constr;
   embed = constr2lp;
 }
 let failsafe_term = {
   CConv.ty = Conv.TyName "term";
   pp_doc = (fun fmt () -> Format.fprintf fmt "A Coq term containing evars");
-  pp = (fun fmt t -> Format.fprintf fmt "%s" (Pp.string_of_ppcmds (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
+  pp = (fun fmt t -> Format.fprintf fmt "@[%a@]" Pp.pp_with ( (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
   readback = (fun ~depth coq_ctx csts s t -> lp2constr ~depth { coq_ctx with options = { coq_ctx.options with failsafe = true }} csts s t);
   embed = constr2lp;
 }
@@ -192,7 +264,7 @@ let proof_context : (full coq_context, API.Data.constraints) CConv.ctx_readback 
 let closed_term = {
   CConv.ty = Conv.TyName "term";
   pp_doc = (fun fmt () -> Format.fprintf fmt "A closed Coq term");
-  pp = (fun fmt t -> Format.fprintf fmt "%s" (Pp.string_of_ppcmds (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
+  pp = (fun fmt t -> Format.fprintf fmt "@[%a@]" Pp.pp_with ( (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
   readback = lp2constr_closed;
   embed = constr2lp_closed
 }
@@ -204,7 +276,7 @@ let global : (empty coq_context, API.Data.constraints) CConv.ctx_readback =
 let closed_ground_term = {
   CConv.ty = Conv.TyName "term";
   pp_doc = (fun fmt () -> Format.fprintf fmt "A ground, closed, Coq term");
-  pp = (fun fmt t -> Format.fprintf fmt "%s" (Pp.string_of_ppcmds (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
+  pp = (fun fmt t -> Format.fprintf fmt "@[%a@]" Pp.pp_with ( (Printer.pr_econstr_env (Global.env()) Evd.empty t)));
   readback = lp2constr_closed_ground;
   embed = constr2lp_closed_ground
 }
@@ -215,7 +287,7 @@ let term_skeleton =  {
   pp = (fun fmt t ->
       let env = Global.env() in
       let sigma = Evd.from_env env in
-      Format.fprintf fmt "%s" (Pp.string_of_ppcmds (Printer.pr_glob_constr_env env sigma t)));
+      Format.fprintf fmt "@[%a@]" Pp.pp_with ( (Printer.pr_glob_constr_env env sigma t)));
   readback = lp2skeleton;
   embed = (fun ~depth _ _ _ _ -> assert false);
 }
@@ -234,7 +306,7 @@ let id = { B.string with
 
 let flag name = { (unspec bool) with Conv.ty = Conv.TyName name }
 
-(* Unfortunately the data tye is not symmeteric *)
+(* Unfortunately the data type is not symmeteric *)
 let indt_decl_in = {
   CConv.ty = Conv.TyName "indt-decl";
   pp_doc = (fun fmt () -> Format.fprintf fmt "Declaration of an inductive type");
@@ -355,7 +427,7 @@ let cs_pattern =
 let cs_instance = let open Conv in let open API.AlgebraicData in let open Recordops in declare {
   ty = TyName "cs-instance";
   doc = "Canonical Structure instances: (cs-instance Proj ValPat Inst)";
-  pp = (fun fmt (_,{ o_DEF }) -> Format.fprintf fmt "%s" Pp.(string_of_ppcmds (Printer.pr_constr_env (Global.env()) Evd.empty o_DEF)));
+  pp = (fun fmt (_,{ o_DEF }) -> Format.fprintf fmt "@[%a@]" Pp.pp_with ((Printer.pr_constr_env (Global.env()) Evd.empty o_DEF)));
   constructors = [
     K("cs-instance","",A(gref,A(cs_pattern,CA(closed_ground_term,N))), (* XXX should be a gref *)
       B (fun p v i -> assert false),
@@ -431,7 +503,7 @@ The name and the grafting specification can be left unspecified.|};
 } |> CConv.(!<)
 
 let set_accumulate_to_db, get_accumulate_to_db =
-  let f = ref ((fun () -> assert false),(fun _ _ ~local:_ -> assert false),(fun () -> assert false)) in
+  let f = ref (fun _ _ ~local:_ -> assert false) in
   (fun x -> f := x),
   (fun () -> !f)
 
@@ -608,6 +680,136 @@ let gr2path state gr =
         ((mp2path @@ MutInd.modpath i) @ [klbl])
     | Names.GlobRef.IndRef _  | Names.GlobRef.ConstructRef _ ->
           nYI "mutual inductive (make-derived...)"
+
+let ppbox = let open Conv in let open Pp in let open API.AlgebraicData in declare {
+  ty = TyName "coq.pp.box";
+  doc = {|Coq box types for pretty printing:
+- Vertical block: each break leads to a new line
+- Horizontal block: no line breaking
+- Horizontal-vertical block: same as Vertical block, except if this block
+  is small enough to fit on a single line in which case it is the same
+  as a Horizontal block
+- Horizontal or Vertical block: breaks lead to new line only when
+  necessary to print the content of the block (the contents flow
+  inside the box)|};
+  pp = (fun fmt _ -> Format.fprintf fmt "<todo>");
+  constructors = [
+    K("coq.pp.v","",A(B.int,N),
+        B (fun i -> Pp_vbox i),
+        M (fun ~ok ~ko -> function Pp_vbox i -> ok i | _ -> ko ()));
+    K("coq.pp.h","",N,
+        B Pp_hbox,
+        M (fun ~ok ~ko -> function Pp_hbox -> ok | _ -> ko ()));
+    K("coq.pp.hv","",A(B.int,N),
+        B (fun i -> Pp_hvbox i),
+        M (fun ~ok ~ko -> function Pp_hvbox i -> ok i | _ -> ko ()));
+    K("coq.pp.hov","",A(B.int,N),
+        B (fun i -> Pp_hovbox i),
+        M (fun ~ok ~ko -> function Pp_hovbox i -> ok i | _ -> ko ()));
+ ]
+} |> CConv.(!<)
+
+let ppboxes = let open Conv in let open Pp in let open API.AlgebraicData in declare {
+  ty = TyName "coq.pp";
+  doc = {|Coq box model for pretty printing. Items:
+- empty
+- spc: a spacem, also a breaking hint
+- str: a non breakable string
+- brk L I: a breaking hint of a given length L contributing I spaces to
+  indentation when taken
+- glue: puts things together
+- box B: a box with automatic line breaking according to B
+- comment: embedded \\n are turned into nl (see below)
+- tag: ignored
+- nl: break the line (should not be used)|};
+  pp = (fun fmt _ -> Format.fprintf fmt "<todo>");
+  constructors = [
+    K("coq.pp.empty","",N,
+        B Ppcmd_empty,
+        M (fun ~ok ~ko -> function Ppcmd_empty -> ok | _ -> ko ()));
+    K("coq.pp.spc","",N,
+        B (Ppcmd_print_break(1,0)),
+        M (fun ~ok ~ko -> function Ppcmd_print_break(1,0) -> ok | _ -> ko ()));
+    K("coq.pp.str","",A(B.string,N),
+        B (fun s -> Ppcmd_string s),
+        M (fun ~ok ~ko -> function Ppcmd_string s -> ok s | _ -> ko ()));
+    K("coq.pp.brk","",A(B.int,A(B.int,N)),
+        B (fun i j -> Ppcmd_print_break(i,j)),
+        M (fun ~ok ~ko -> function Ppcmd_print_break(i,j) -> ok i j | _ -> ko ()));
+    K("coq.pp.glue","",C((fun pp -> CConv.(!>>) B.list pp),N),
+        B (fun l -> Ppcmd_glue (List.map Pp.unrepr l)),
+        M (fun ~ok ~ko -> function Ppcmd_glue l -> ok (List.map Pp.repr l) | _ -> ko ()));
+    K("coq.pp.box","",A(ppbox,C((fun pp -> CConv.(!>>) B.list pp),N)),
+        B (fun b l -> Ppcmd_box(b,Pp.unrepr @@ Ppcmd_glue (List.map Pp.unrepr l))),
+        M (fun ~ok ~ko -> function
+          | Ppcmd_box(b,x) ->
+            begin match Pp.repr x with
+            | Ppcmd_glue l -> ok b (List.map Pp.repr l)
+            | x -> ok b [x]
+            end
+          | _ -> ko ()));
+    K("coq.pp.comment","",A(B.list B.string,N),
+        B (fun l -> Ppcmd_comment l),
+        M (fun ~ok ~ko -> function Ppcmd_comment l -> ok l | _ -> ko ()));
+    K("coq.pp.tag","",A(B.string,S N),
+        B (fun b x -> Ppcmd_tag(b,Pp.unrepr x)),
+        M (fun ~ok ~ko -> function Ppcmd_tag(b,x) -> ok b (Pp.repr x) | _ -> ko ()));
+    K("coq.pp.nl","",N,
+        B Ppcmd_force_newline,
+        M (fun ~ok ~ko -> function Ppcmd_force_newline -> ok | _ -> ko ()));
+  ]
+} |> CConv.(!<)
+
+let warn_deprecated_add_axiom =
+  CWarnings.create
+    ~name:"elpi.add-const-for-axiom-or-sectionvar" 
+    ~category:"deprecated"
+    Pp.(fun () ->
+         strbrk ("elpi: Using coq.env.add-const for declaring axioms or " ^
+           "section variables is deprecated. Use coq.env.add-axiom or " ^ 
+           "coq.env.add-section-variable instead"))
+         
+let add_axiom_or_variable api id sigma ty local =
+  let used = EConstr.universes_of_constr sigma ty in
+  let sigma = Evd.restrict_universe_context sigma used in
+  let ubinders = Evd.universe_binders sigma in
+  let uentry = Evd.univ_entry ~poly:false sigma in
+  let kind = Decls.Logical in
+  let impargs = [] in
+  let variable = CAst.(make @@ Id.of_string id) in
+  if not (is_ground sigma ty) then
+    err Pp.(str"coq.env.add-const: the type must be ground. Did you forge to call coq.typecheck-indt-decl?");
+  let gr, _ =
+    if local then begin
+      let uctx =
+        let context_set_of_entry = function
+          | Entries.Polymorphic_entry (_,uctx) -> Univ.ContextSet.of_context uctx
+          | Entries.Monomorphic_entry uctx -> uctx in
+        context_set_of_entry uentry in
+      DeclareUctx.declare_universe_context ~poly:false uctx;
+      ComAssumption.declare_variable false ~kind (EConstr.to_constr sigma ty) impargs Glob_term.Explicit variable;
+      GlobRef.VarRef(Id.of_string id), Univ.Instance.empty
+    end else
+      ComAssumption.declare_axiom false ~local:Locality.ImportDefaultBehavior ~poly:false ~kind (EConstr.to_constr sigma ty)
+        (uentry, ubinders) impargs Declaremods.NoInline
+        variable
+  in
+  gr
+  ;;
+
+
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+(*****************************************************************************)
+
 
 let coq_builtins =
   let open API.BuiltIn in
@@ -973,7 +1175,8 @@ and the current module; Ty can be left unspecified and in that case the
 inferred one is taken (as in writing Definition x := t); Bo can be left
 unspecified and in that case an axiom is added (or a section variable,
 if a section is open and @local! is used). Omitting the body and the type is
-an error.
+an error. Note: using this API for declaring an axiom or a section variable is
+deprecated, use coq.env.add-axiom or coq.env.add-section-variable instead.
 Supported attributes:
 - @local! (default: false)|})))))),
   (fun id body types opaque _ ~depth {options} _ -> on_global_state "coq.env.add-const" (fun state ->
@@ -985,30 +1188,8 @@ Supported attributes:
        | Unspec ->
          err Pp.(str "coq.env.add-const: both Type and Body are unspecified")
        | Given ty ->
-       let used = EConstr.universes_of_constr sigma ty in
-       let sigma = Evd.restrict_universe_context sigma used in
-       let ubinders = Evd.universe_binders sigma in
-       let uentry = Evd.univ_entry ~poly:false sigma in
-       let kind = Decls.Logical in
-       let impargs = [] in
-       let variable = CAst.(make @@ Id.of_string id) in
-       if not (is_ground sigma ty) then
-         err Pp.(str"coq.env.add-const: the type must be ground. Did you forge to call coq.typecheck-indt-decl?");
-       let gr, _ =
-         if local then begin
-           let uctx =
-              let context_set_of_entry = function
-                | Entries.Polymorphic_entry (_,uctx) -> Univ.ContextSet.of_context uctx
-                | Entries.Monomorphic_entry uctx -> uctx in
-              context_set_of_entry uentry in
-           DeclareUctx.declare_universe_context ~poly:false uctx;
-           ComAssumption.declare_variable false ~kind (EConstr.to_constr sigma ty) impargs Glob_term.Explicit variable;
-           GlobRef.VarRef(Id.of_string id), Univ.Instance.empty
-         end else
-           ComAssumption.declare_axiom false ~local:Locality.ImportDefaultBehavior ~poly:false ~kind (EConstr.to_constr sigma ty)
-             (uentry, ubinders) impargs Declaremods.NoInline
-             variable
-       in
+       warn_deprecated_add_axiom ();
+       let gr = add_axiom_or_variable "coq.env.add-const" id sigma ty local in
        state, !: (global_constant_of_globref gr), []
      end
     | Given body ->
@@ -1030,6 +1211,30 @@ Supported attributes:
        let info = Declare.Info.make ~scope ~kind ~poly:false ~udecl () in
        let gr = Declare.declare_definition ~cinfo ~info ~opaque:(opaque = Given true) ~body sigma in
        state, !: (global_constant_of_globref gr), []))),
+  DocAbove);
+
+  MLCode(Pred("coq.env.add-axiom",
+    In(id,   "Name",
+    CIn(closed_ground_term, "Ty",
+    Out(constant, "C",
+    Full (global, {|Declare a new axiom: C gets a constant derived from Name
+and the current module|})))),
+  (fun id ty _ ~depth _ _ -> on_global_state_does_rewind_env "coq.env.add-axiom" (fun state ->
+     let sigma = get_sigma state in
+     let gr = add_axiom_or_variable "coq.env.add-axiom" id sigma ty false in
+     state, !: (global_constant_of_globref gr), []))),
+  DocAbove);
+
+  MLCode(Pred("coq.env.add-section-variable",
+    In(id,   "Name",
+    CIn(closed_ground_term, "Ty",
+    Out(constant, "C",
+    Full (global, {|Declare a new section variable: C gets a constant derived from Name
+and the current module|})))),
+  (fun id ty _ ~depth _ _ -> on_global_state_does_rewind_env "coq.env.add-section-variable" (fun state ->
+     let sigma = get_sigma state in
+     let gr = add_axiom_or_variable "coq.env.add-section-variable" id sigma ty true in
+     state, !: (global_constant_of_globref gr), []))),
   DocAbove);
 
   MLCode(Pred("coq.env.add-indt",
@@ -1709,7 +1914,7 @@ Universe constraints are put in the constraint store.|})))),
           (* optimization: don't print the error if caller wants OK *)
           raise No_clause
        | _ ->
-          let error = Pp.string_of_ppcmds @@ Himsg.explain_pretype_error env sigma err in
+          let error = string_of_ppcmds proof_context.options @@ Himsg.explain_pretype_error env sigma err in
           state, ?: None +! B.mkERROR error, [])),
   DocAbove);
 
@@ -1741,7 +1946,7 @@ Universe constraints are put in the constraint store.|})))),
           (* optimization: don't print the error if caller wants OK *)
           raise No_clause
        | _ ->
-          let error = Pp.string_of_ppcmds @@ Himsg.explain_pretype_error env sigma err in
+          let error = string_of_ppcmds proof_context.options @@ Himsg.explain_pretype_error env sigma err in
           state, ?: None +! B.mkERROR error, [])),
   DocAbove);
 
@@ -1762,7 +1967,7 @@ Universe constraints are put in the constraint store.|})))),
           (* optimization: don't print the error if caller wants OK *)
           raise No_clause
        | _ ->
-          let error = Pp.string_of_ppcmds @@ Himsg.explain_pretype_error env sigma err in
+          let error = string_of_ppcmds proof_context.options @@ Himsg.explain_pretype_error env sigma err in
           state, !: (B.mkERROR error), [])),
   DocAbove);
 
@@ -1783,7 +1988,7 @@ Universe constraints are put in the constraint store.|})))),
           (* optimization: don't print the error if caller wants OK *)
           raise No_clause
        | _ ->
-          let error = Pp.string_of_ppcmds @@ Himsg.explain_pretype_error env sigma err in
+          let error = string_of_ppcmds proof_context.options @@ Himsg.explain_pretype_error env sigma err in
           state, !: (B.mkERROR error), [])),
   DocAbove);
 
@@ -1816,7 +2021,7 @@ hole. Similarly universe levels present in T are disregarded.|}))))),
           (* optimization: don't print the error if caller wants OK *)
           raise No_clause
        | _ ->
-          let error = Pp.string_of_ppcmds @@ Himsg.explain_pretype_error env sigma err in
+          let error = string_of_ppcmds proof_context.options @@ Himsg.explain_pretype_error env sigma err in
           state, ?: None +? None +! B.mkERROR error, [])),
   DocAbove);
 
@@ -1844,7 +2049,7 @@ hole. Similarly universe levels present in T are disregarded.|}))))),
           (* optimization: don't print the error if caller wants OK *)
           raise No_clause
        | _ ->
-          let error = Pp.string_of_ppcmds @@ Himsg.explain_pretype_error env sigma err in
+          let error = string_of_ppcmds proof_context.options @@ Himsg.explain_pretype_error env sigma err in
           state, ?: None +? None +! B.mkERROR error, [])),
   DocAbove);
 
@@ -2034,10 +2239,29 @@ coq.id->name S N :- coq.string->name S N.
   MLCode(Pred("coq.term->string",
     CIn(failsafe_term,"T",
     Out(B.string, "S",
-    Full(proof_context, "prints a term T to a string S using Coq's pretty printer"))),
+    Full(proof_context, {|prints a term T to a string S using Coq's pretty printer
+Supported attributes:
+- @ppwidth! N (default 80, max line length)
+- @ppall! (default: false, prints all details)
+- @ppmost! (default: false, prints most details)
+- @holes! (default: false, prints evars as _)|}))),
   (fun t _ ~depth proof_context constraints state ->
      let sigma = get_sigma state in
-     let s = Pp.string_of_ppcmds (Printer.pr_econstr_env proof_context.env sigma t) in
+     let s = string_of_ppcmds proof_context.options (pr_econstr_env proof_context.options proof_context.env sigma t) in
+     state, !: s, [])),
+  DocAbove);
+
+  MLCode(Pred("coq.term->pp",
+    CIn(failsafe_term,"T",
+    Out(ppboxes, "B",
+    Full(proof_context, {|prints a term T to a pp.t B using Coq's pretty printer"
+Supported attributes:
+- @ppall! (default: false, prints all details)
+- @ppmost! (default: false, prints most details)
+- @holes! (default: false, prints evars as _)|}))),
+  (fun t _ ~depth proof_context constraints state ->
+     let sigma = get_sigma state in
+     let s = Pp.repr @@ pr_econstr_env proof_context.options proof_context.env sigma t in
      state, !: s, [])),
   DocAbove);
 
@@ -2069,8 +2293,8 @@ Supported attributes:
          State.update clauses_for_later state (fun l ->
            (dbname,clause,local) :: l), (), []
      | Given CurrentModule ->
-          let elpi, f, cur_program = get_accumulate_to_db () in
-          f dbname API.(Compile.unit ~follows:(cur_program ()) ~elpi:(elpi ()) ~flags:Compile.default_flags clause) ~local;
+          let f = get_accumulate_to_db () in
+          f dbname clause ~local;
           state, (), []
      )),
   DocAbove);
@@ -2078,6 +2302,21 @@ Supported attributes:
   LPDoc "-- Utils ------------------------------------------------------------";
   ] @
   B.ocaml_set ~name:"coq.gref.set" gref (module GRSet) @
-  B.ocaml_map ~name:"coq.gref.map" gref (module GRMap)
+  B.ocaml_map ~name:"coq.gref.map" gref (module GRMap) @
+  [
+  MLData ppbox;
+  MLData ppboxes;
+  MLCode(Pred("coq.pp->string",
+    In(ppboxes, "B",
+    Out(B.string, "S",
+    Read(global, {|Prints a pp.t box expression B to a string S
+Supported attributes:
+- @ppwidth! N (default 80, max line length)|}))),
+  (fun box _ ~depth ctx _ _ ->
+     !: (string_of_ppcmds ctx.options (Pp.unrepr box))
+     )),
+  DocAbove)
+
+  ]
 
 ;;