+++ src/coq_elpi_vernacular.ml
@@ -14,15 +14,37 @@ let debug_vars = Summary.ref ~name:"elpi-debug" EC.StrSet.empty
 let cc_flags () =
   { EC.default_flags with EC.defined_variables = !debug_vars }
 
+let source_cache = Summary.Local.ref
+  ~name:"elpi-source-cache"
+  CString.Map.empty
+
+(* Source files can be large, and loaded multiple times since many entry point
+   can be implemented in the same file. We share (in memory) the parsed file. *)
 let unit_from_file ~elpi x =
-  try EC.unit ~elpi ~flags:(cc_flags ()) (EP.program ~elpi ~print_accumulated_files:false [x])
-  with
-  | EP.ParseError(loc, msg) ->
-    let loc = Coq_elpi_utils.to_coq_loc loc in
-    CErrors.user_err ~loc ~hdr:"elpi" (Pp.str msg)
-  | EC.CompileError(oloc, msg) ->
-    let loc = Option.map Coq_elpi_utils.to_coq_loc oloc in
-    CErrors.user_err ?loc ~hdr:"elpi" (Pp.str msg)
+  let open Summary.Local in
+  let flags = cc_flags () in
+  try
+    let hash = Digest.file (EP.resolve_file x) in
+    let u, old_flags, old_hash = CString.Map.find x !source_cache in
+    if flags <> old_flags then raise Not_found;
+    if hash <> old_hash then raise Not_found;
+    u
+  with Not_found ->
+    try
+      let hash = Digest.file (EP.resolve_file x) in
+      let program = EP.program ~elpi ~print_accumulated_files:false [x] in
+      let u = EC.unit ~elpi ~flags program in
+      source_cache := CString.Map.add x (u,flags,hash) !source_cache;
+      u
+    with
+    | Sys_error msg ->
+      CErrors.user_err ~hdr:"elpi" (Pp.str msg)
+    | EP.ParseError(loc, msg) ->
+      let loc = Coq_elpi_utils.to_coq_loc loc in
+      CErrors.user_err ~loc ~hdr:"elpi" (Pp.str msg)
+    | EC.CompileError(oloc, msg) ->
+      let loc = Option.map Coq_elpi_utils.to_coq_loc oloc in
+      CErrors.user_err ?loc ~hdr:"elpi" (Pp.str msg)
 
 let unit_from_string ~elpi loc x =
   let x = Stream.of_string x in
@@ -42,13 +64,13 @@ let parse_goal loc x =
     CErrors.user_err ~loc ~hdr:"elpi" (Pp.str msg)
 
 let assemble_units ~elpi units =
-  try EC.assemble ~elpi units
+  try EC.assemble ~elpi ~flags:(cc_flags ()) units
   with EC.CompileError(oloc, msg) ->
     let loc = Option.map Coq_elpi_utils.to_coq_loc oloc in
     CErrors.user_err ?loc ~hdr:"elpi" (Pp.str msg)
 
 let extend_w_units ~base units =
-  try EC.extend ~base units
+  try EC.extend ~flags:(cc_flags ()) ~base units
   with EC.CompileError(oloc, msg) ->
     let loc = Option.map Coq_elpi_utils.to_coq_loc oloc in
     CErrors.user_err ?loc ~hdr:"elpi" (Pp.str msg)
@@ -303,7 +325,7 @@ and src_string = {
   val set_current_program : qualified_name -> unit
   val current_program : unit -> qualified_name
   val accumulate : qualified_name -> src list -> unit
-  val accumulate_to_db : qualified_name -> Compile.compilation_unit list -> local:bool -> unit
+  val accumulate_to_db : qualified_name -> Compile.compilation_unit -> local:bool -> unit
 
   val load_checker : string -> unit
   val load_printer : string -> unit
@@ -347,19 +369,23 @@ let program_src : (SrcSet.t * src list) SLMap.t ref =
 let program_exists name = SLMap.mem name !program_src
 
 module KSet = Set.Make(Names.KerName)
-let db_name_src : (KSet.t * (Names.KerName.t * EC.compilation_unit list) list) SLMap.t ref =
+type db = {
+  base : EC.compilation_unit;
+  extensions : KSet.t;
+  contents : (Names.KerName.t * EC.compilation_unit) list;
+}
+
+let db_name_src : db SLMap.t ref =
   Summary.ref ~name:"elpi-db" SLMap.empty
 let db_exists name = SLMap.mem name !db_name_src
 
 let ast_of_src = function
-  | File { fast = a } -> [`Static,a]
-  | EmbeddedString { sast = a } -> [`Static,a]
+  | File { fast = a } -> [`Static,(a)]
+  | EmbeddedString { sast = a } -> [`Static,(a)]
   | Database name ->
      try
-       match List.(flatten (map snd (snd (SLMap.find name !db_name_src)))) with
-       | header :: clauses ->
-           (`Static,header) :: List.map (fun x -> `Dynamic,x) clauses
-       | [] -> assert false
+       let { base; contents } = SLMap.find name !db_name_src in
+        (`Static,(base)) :: List.map (fun (_,x) -> `Dynamic,x) contents
      with Not_found ->
        CErrors.user_err Pp.(str "Unknown Db " ++ str (show_qualified_name name))
 
@@ -440,15 +466,19 @@ let add_to_program name v =
 ;;
 
 let append_to_db name (kname,data as l) =
-  let seen, db = try SLMap.find name !db_name_src with Not_found -> KSet.empty, [] in
-  if KSet.mem kname seen then seen, db
-  else
-    let seen = KSet.add kname seen in
-    seen, db @ [l]
+  try
+      let db = SLMap.find name !db_name_src in
+      if KSet.mem kname db.extensions then db
+      else { db with contents = db.contents @ [l];
+                    extensions = KSet.add kname db.extensions }
+  with Not_found ->
+    (* We assume the first chunk is the base *)
+    let extensions = KSet.singleton kname in
+    { base = data; extensions; contents = [];}
 
 type snippet = {
   program : qualified_name;
-  code : EC.compilation_unit list;
+  code : EC.compilation_unit;
   local : bool;
 }
 
@@ -464,8 +494,9 @@ let in_db : snippet -> Libobject.obj =
     Libobject.open_function = Libobject.simple_open import;
 }
 
+let accum = ref 0
 let add_to_db program code local =
-  Lib.add_anonymous_leaf (in_db { program; code; local })
+  ignore @@ Lib.add_leaf (Names.Id.of_string (incr accum; Printf.sprintf "_ELPI_%d" !accum)) (in_db { program; code; local })
 
 let lp_command_ast = Summary.ref ~name:"elpi-lp-command" None
 let in_lp_command_src : src -> Libobject.obj =
@@ -513,7 +544,9 @@ let create_db (loc,qualid) init =
   if program_exists qualid || db_exists qualid then
     CErrors.user_err Pp.(str "Program/Tactic/Db " ++ pr_qualified_name qualid ++ str " already exists")
   else
-    add_to_db qualid (List.map snd @@ ast_of_src init) false
+    match ast_of_src init with
+    | [`Static,base] -> add_to_db qualid base false
+    | _ -> assert false
 ;;
 
 let set_current_program n =
@@ -578,9 +611,6 @@ let accumulate_to_db p v ~local =
 
 end
 
-let current_program_begin_run = ref None
-
-let () = Coq_elpi_builtins.set_accumulate_to_db (Programs.ensure_initialized, (fun n x ~local -> Programs.accumulate_to_db n [x] ~local), (fun () -> Option.get !current_program_begin_run))
 open Programs
 
 let load_command = load_command
@@ -638,6 +668,15 @@ let compile name baseul extra =
     (Summary.Local.(:=) compiler_cache (SLMap.add name (baseul,program) (Summary.Local.(!) compiler_cache)));
     extend_w_units ~base:program extra
 
+let () = Coq_elpi_builtins.set_accumulate_to_db (fun n x ~local ->
+  let elpi = ensure_initialized () in
+  let u =
+    try EC.unit ~elpi ~flags:(cc_flags ()) x
+    with EC.CompileError(oloc, msg) ->
+      let loc = Option.map Coq_elpi_utils.to_coq_loc oloc in
+      CErrors.user_err ?loc ~hdr:"elpi" (Pp.str msg) in
+  Programs.accumulate_to_db n u ~local)
+
 let get_and_compile name =
   let core_units, extra_units = get name in
   let prog = compile name core_units extra_units in
@@ -649,7 +688,6 @@ let run_static_check query =
   ignore (EC.static_check ~checker query)
 
 let run ~tactic_mode ~static_check program query =
-  current_program_begin_run := Some program;
   let t1 = Unix.gettimeofday () in
   let query =
     match query with
@@ -707,7 +745,7 @@ let run_and_print ~tactic_mode ~print ~static_check program_ast query_ast =
     let elpi = ensure_initialized () in
     List.iter (fun (dbname,ast,local) ->
       let unit = EC.unit ~elpi ~flags:(cc_flags()) ast in
-      accumulate_to_db dbname [unit] ~local) clauses_to_add
+      accumulate_to_db dbname unit ~local) clauses_to_add
 ;;
 
 let run_in_program ?(program = current_program ()) (loc, query) =
@@ -741,18 +779,29 @@ let mainc = ET.Constants.declare_global_symbol "main"
 let attributesc = ET.Constants.declare_global_symbol "attributes"
 
 let run_program loc name ~atts args =
+  let loc = Coq_elpi_utils.of_coq_loc loc in
   let args = args
     |> List.map (glob_arg (Genintern.empty_glob_sign (Global.env())))
     |> List.map (interp_arg (Ltac_plugin.Tacinterp.default_ist ()) Evd.({ sigma = from_env (Global.env()); it = 0 }))
     |> List.map snd in
   let args = args |> List.map to_arg in
   let query ~depth state =
+    let atts = ("elpi.loc", Attributes.(VernacFlagLeaf (FlagString (API.Ast.Loc.show loc)))) :: atts in
+    let atts =
+      match Sys.getenv_opt "COQ_ELPI_ATTRIBUTES" with
+      | None -> atts
+      | Some txt ->
+          match Pcoq.parse_string (Pvernac.main_entry None) (Printf.sprintf "#[%s] Qed." txt) |> Option.map (fun x -> x.CAst.v) with
+          | None -> atts
+          | Some { Vernacexpr.attrs ; _ } -> List.map (fun (name,v) -> "elpi."^name,v) attrs @ atts
+          | exception Stream.Error msg ->
+              CErrors.user_err ~hdr:"elpi" Pp.(str"Environment variable COQ_ELPI_ATTRIBUTES contains ill formed value:" ++ spc () ++ str txt ++ cut () ++ str msg) in
     let state, atts, _ = EU.map_acc (Coq_elpi_builtins.attribute.API.Conversion.embed ~depth) state atts in
     let state, args = CList.fold_left_map
       (Coq_elpi_goal_HOAS.in_elpi_global_arg ~depth Coq_elpi_HOAS.(mk_coq_context ~options:default_options state))
       state args in
     let atts = ET.mkApp attributesc (EU.list_to_lp_list atts) [] in
-    state, (Coq_elpi_utils.of_coq_loc loc, ET.mkApp ET.Constants.implc atts [ET.mkApp mainc (EU.list_to_lp_list args) []]) in
+    state, (loc, ET.mkApp ET.Constants.implc atts [ET.mkApp mainc (EU.list_to_lp_list args) []]) in
   let program = get_and_compile name in
   run_and_print ~tactic_mode:false ~print:false ~static_check:false program (`Fun query)
 ;;
@@ -835,7 +884,6 @@ let run_in_tactic ?(program = current_program ()) (loc,query) ist args =
   | _ -> tclZEROMSG Pp.(str "elpi fails")
 end end end
 
-
 let accumulate_files ?(program=current_program()) s =
   let elpi = ensure_initialized () in
   try
@@ -852,7 +900,7 @@ let accumulate_string ?(program=current_program()) (loc,s) =
   let elpi = ensure_initialized () in
   let new_ast = unit_from_string ~elpi loc s in
   if db_exists program then
-    accumulate_to_db program [new_ast] ~local:false
+    accumulate_to_db program new_ast ~local:false
   else
     accumulate program [EmbeddedString { sloc = loc; sdata = s; sast = new_ast}]
 ;;
@@ -866,10 +914,9 @@ let accumulate_db ?(program=current_program()) name =
 let accumulate_to_db db (loc,s) =
   let elpi = ensure_initialized () in
   let new_ast = unit_from_string ~elpi loc s in
-  if db_exists db then accumulate_to_db db [new_ast]
+  if db_exists db then accumulate_to_db db new_ast
   else CErrors.user_err
     Pp.(str "Db " ++ pr_qualified_name db ++ str" not found") 
-
 
 let in_exported_program : (qualified_name * string * (Loc.t,Loc.t,Loc.t) Genarg.ArgT.tag * (raw_arg,glob_arg,parsed_arg) Genarg.ArgT.tag) -> Libobject.obj =
   Libobject.declare_object @@ Libobject.global_object_nodischarge "ELPI-EXPORTED"
@@ -879,12 +926,13 @@ let in_exported_program : (qualified_name * string * (Loc.t,Loc.t,Loc.t) Genarg.
       ~classifier:(fun _ -> Vernacextend.classify_as_sideeff)
       ?entry:None
       [ Vernacextend.TyML (false,
-          Vernacextend.TyTerminal (p_str,
-            Vernacextend.TyNonTerminal (Extend.TUentry tag_loc,
-            Vernacextend.TyNonTerminal (Extend.TUlist0 (Extend.TUentry tag_arg),
-            Vernacextend.TyNil))),
-          (fun loc args ~atts -> Vernacextend.VtDefault (fun () ->
-              run_program loc p ~atts args)),
+         Vernacextend.TyNonTerminal (Extend.TUentry tag_loc,
+         Vernacextend.TyTerminal (p_str,
+         Vernacextend.TyNonTerminal (Extend.TUlist0 (Extend.TUentry tag_arg),
+         Vernacextend.TyNonTerminal (Extend.TUentry tag_loc,
+         Vernacextend.TyNil)))),
+          (fun loc0 args loc1 (* 8.14 ~loc*) ~atts -> Vernacextend.VtDefault (fun () ->
+              run_program (Loc.merge loc0 loc1) (*loc*) p ~atts args)),
           None)])
     ~subst:(Some (fun _ -> CErrors.user_err Pp.(str"elpi: No functors yet")))